LPS-31119 Fix non compatible SQL query
All the fields named in the select list must be named in the group by clause. For fields which may have multiple values in the same group, a single value criteria must be forced. Thus, for kbArticleId, userId, userName and status (which may have multiple values for the same resourcePrimKey, groupId, companyId group) the MIN function is used to select a single value.